backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
motor==3.5.1
pymongo==4.8.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
python-dotenv==1.0.0
python-jwt==3.3.4
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.25.2
APScheduler==3.10.4